whitespaces

+++ skarlygina.mariia/I4/count_diagonals_without_zeros.h
@@ -1,5 +1,5 @@
 #ifndef COUNT_DIAGONALS_WITHOUT_ZEROS_H
 #define COUNT_DIAGONALS_WITHOUT_ZEROS_H
 #include <cstddef>
-size_t countDiagonalsWithoutZeros(const int** array, size_t rows, size_t columns);
+unsigned int countDiagonalsWithoutZeros(const int** array, size_t rows, size_t columns);
 #endif
